Extract REST multipart tests in a separate file
