+++ tests/async_bridge/mod.rs
@@ -1,2 +1,3 @@
 mod graphql;
 mod rest;
+mod rest_multipart;
